feat(sidebar): 响应式折叠 — 桌面三态切换 + 移动抽屉
- 桌面端(≥768px)三态：展开(14rem)↔图标条(3.5rem)↔隐藏(0)
  单按钮循环切换，localStorage 持久化
  图标条态：导航仅 icon + CSS tooltip，指数轮播，RadioTower 实时行情按钮
- 移动端(<768px)抽屉模式：FAB 汉堡按钮，80vw/max 320px
  点遮罩/选导航/ESC 关闭，body 滚动锁
- Layout.tsx 从 698 行瘦身为 80 行壳
  侧边栏内容下沉到 components/sidebar/ 8 个文件
  SSE 订阅/断线提示/监控徽标轮询保留在 Layout
- 共享逻辑抽取：useVisibleNavItems/useRealtimeToggle/useDataSyncStatus
- .gitignore 忽略 docs/superpowers/ 和 .codebuddy.md
