+++ frontend/src/components/sidebar/components.tsx
@@ -1,0 +1,206 @@
+import { NavLink } from 'react-router-dom'
+import { Key, Settings, Sparkles, Sun, Moon } from 'lucide-react'
+import { toggleTheme, useTheme } from '@/lib/theme'
+import { useUnreadAlerts } from '@/lib/monitorBadge'
+import {
+  BRAND,
+  fmtIndexValue,
+  fmtIndexPct,
+  indexPctClass,
+  type CoreIndex,
+} from './shared'
+import type { IndexQuote } from '@/lib/api'
+
+/** 亮/暗主题切换 */
+export function ThemeToggle() {
+  const theme = useTheme()
+  const dark = theme === 'dark'
+  return (
+    <button
+      onClick={() => toggleTheme()}
+      className="flex items-center justify-center rounded-btn p-2 text-foreground/80 transition-colors duration-150 ease-smooth hover:bg-elevated hover:text-foreground cursor-pointer"
+      title={dark ? '切换到亮色模式' : '切换到暗色模式'}
+    >
+      {dark ? <Sun className="h-4 w-4 shrink-0" /> : <Moon className="h-4 w-4 shrink-0" />}
+    </button>
+  )
+}
+
+/** 档位卡片 — 展开态用 */
+export function TierBadge({ label, hasKey }: { label: string; hasKey?: boolean }) {
+  const base = label.split(' ')[0].split('+')[0].toLowerCase()
+  const isNone = base === 'none'
+
+  const tierConfig: Record<string, {
+    desc: string
+    tagBg: React.CSSProperties
+    dotStyle: React.CSSProperties
+    labelTextStyle: React.CSSProperties
+  }> = {
+    none: {
+      desc: '未配置 Key · 仅历史日K',
+      tagBg: { background: 'rgba(113,113,122,0.15)' },
+      dotStyle: { background: '#52525b' },
+      labelTextStyle: { color: '#71717a' },
+    },
+    free: {
+      desc: '基础日K · 自选实时',
+      tagBg: { background: 'rgba(113,113,122,0.3)' },
+      dotStyle: { background: '#71717a' },
+      labelTextStyle: { color: '#a1a1aa' },
+    },
+    starter: {
+      desc: '批量同步 · 行情池',
+      tagBg: { background: 'rgba(59,130,246,0.2)' },
+      dotStyle: { background: '#3b82f6' },
+      labelTextStyle: { color: '#60a5fa' },
+    },
+    pro: {
+      desc: '分钟K · 实时行情 · 盘口',
+      tagBg: { background: 'linear-gradient(135deg, rgba(168,85,247,0.2), rgba(124,58,237,0.15))' },
+      dotStyle: { background: 'linear-gradient(135deg, #a855f7, #7c3aed)' },
+      labelTextStyle: { background: 'linear-gradient(135deg, #c084fc, #a855f7)', WebkitBackgroundClip: 'text', backgroundClip: 'text', color: 'transparent' },
+    },
+    expert: {
+      desc: 'WebSocket · 财务数据',
+      tagBg: { background: 'linear-gradient(135deg, rgba(59,130,246,0.2), rgba(168,85,247,0.2), rgba(245,158,11,0.2))' },
+      dotStyle: { background: 'linear-gradient(135deg, #3b82f6, #a855f7, #f59e0b)' },
+      labelTextStyle: { background: 'linear-gradient(135deg, #60a5fa, #c084fc, #fbbf24)', WebkitBackgroundClip: 'text', backgroundClip: 'text', color: 'transparent' },
+    },
+  }
+
+  const t = tierConfig[base] || tierConfig.none
+  const displayLabel = isNone ? 'None' : (label || 'None')
+
+  return (
+    <NavLink
+      to="/settings?tab=account"
+      className="mt-2.5 group block -mx-2.5"
+      title="API 设置"
+    >
+      <div className="relative overflow-hidden rounded-lg border border-blue-400/20 bg-gradient-to-br from-blue-500/[0.12] via-surface to-surface px-3 py-2 transition-all hover:border-blue-400/35 hover:from-blue-500/[0.16]">
+        <div className="absolute -right-5 -top-6 h-14 w-14 rounded-full bg-blue-500/10 blur-2xl" />
+        <div className="relative flex items-center gap-2">
+          <div className="flex h-6 w-6 items-center justify-center rounded-md bg-blue-400/10 text-blue-300 ring-1 ring-blue-400/20">
+            <Key className="h-3.5 w-3.5" />
+          </div>
+          <div className="min-w-0 flex-1">
+            <div className="flex items-center gap-1.5">
+              <span className="text-xs font-medium text-foreground">TickFlow</span>
+              <span
+                className="h-1.5 w-1.5 rounded-full"
+                style={{ ...t.dotStyle, ...(base === 'expert' ? { animation: 'pulse 2s infinite' } : {}) }}
+              />
+            </div>
+            <div className="mt-0.5 truncate text-[10px] leading-tight text-muted">
+              {isNone && !hasKey ? '配置 Key 解锁更多能力' : t.desc}
+            </div>
+          </div>
+          <span
+            className="inline-flex h-[18px] max-w-[68px] shrink-0 items-center overflow-hidden rounded px-1.5 text-[10px] font-bold font-mono leading-none"
+            style={t.tagBg}
+          >
+            <span className="truncate" style={t.labelTextStyle}>{displayLabel}</span>
+          </span>
+          <Settings className="h-3 w-3 shrink-0 text-muted group-hover:text-blue-300 transition-colors" />
+        </div>
+      </div>
+    </NavLink>
+  )
+}
+
+/** AI 配置卡片 — 展开态用 */
+export function AIConfigBadge({ configured, model }: { configured?: boolean; model?: string }) {
+  return (
+    <NavLink
+      to="/settings?tab=ai"
+      className="mt-2 group block -mx-2.5"
+      title="AI 配置"
+    >
+      <div className="relative overflow-hidden rounded-lg border border-purple-400/20 bg-gradient-to-br from-purple-500/[0.12] via-surface to-surface px-3 py-2 transition-all hover:border-purple-400/35 hover:from-purple-500/[0.16]">
+        <div className="absolute -right-5 -top-6 h-14 w-14 rounded-full bg-purple-500/10 blur-2xl" />
+        <div className="relative flex items-center gap-2">
+          <div className="flex h-6 w-6 items-center justify-center rounded-md bg-purple-400/10 text-purple-300 ring-1 ring-purple-400/20">
+            <Sparkles className="h-3.5 w-3.5" />
+          </div>
+          <div className="min-w-0 flex-1">
+            <div className="flex items-center gap-1.5">
+              <span className="text-xs font-medium text-foreground">AI 配置</span>
+              <span className={`h-1.5 w-1.5 rounded-full ${configured ? 'bg-bear' : 'bg-warning'}`} />
+            </div>
+            <div className="mt-0.5 truncate text-[10px] leading-tight text-muted">
+              {configured ? (model || '已接入模型') : '接入策略生成模型'}
+            </div>
+          </div>
+          <Settings className="h-3 w-3 text-muted group-hover:text-purple-300 transition-colors" />
+        </div>
+      </div>
+    </NavLink>
+  )
+}
+
+/**
+ * 监控中心未读徽标。
+ * variant="badge" (默认): 带数字的徽标，展开态用
+ * variant="dot": 小红点，图标条态用（absolute 定位）
+ */
+export function MonitorBadge({
+  active,
+  variant = 'badge',
+}: {
+  active: boolean
+  variant?: 'badge' | 'dot'
+}) {
+  const unread = useUnreadAlerts()
+  const badgeEnabled = (() => {
+    try { return localStorage.getItem('monitor_badge_enabled') !== '0' } catch { return true }
+  })()
+  if (active || unread <= 0 || !badgeEnabled) return null
+
+  if (variant === 'dot') {
+    return <span className="absolute -top-0.5 -right-0.5 h-2 w-2 rounded-full bg-danger animate-pulse" />
+  }
+
+  return (
+    <span className="inline-flex h-4 min-w-4 items-center justify-center rounded-full bg-danger px-1 text-[9px] font-bold text-white animate-pulse">
+      {unread > 99 ? '99+' : unread}
+    </span>
+  )
+}
+
+/** 侧边栏指数行情 2x2 网格 — 展开态用 */
+export function SidebarIndexQuotes({
+  rows,
+  items,
+}: {
+  rows: IndexQuote[] | undefined
+  items: CoreIndex[]
+}) {
+  if (items.length === 0) return null
+  const quoteBySymbol = new Map((rows ?? []).map((q) => [q.symbol, q]))
+  return (
+    <div className="mt-2 grid grid-cols-2 gap-1.5">
+      {items.map((item) => {
+        const q = quoteBySymbol.get(item.symbol)
+        const value = q?.last_price ?? q?.close
+        const pct = q?.change_pct
+        return (
+          <NavLink
+            key={item.symbol}
+            to={`/indices?symbol=${encodeURIComponent(item.symbol)}`}
+            className="block rounded bg-elevated/60 px-2 py-1.5 transition-colors hover:bg-elevated"
+            title={`${item.name} ${item.symbol}`}
+          >
+            <div className="flex items-center justify-between gap-1">
+              <span className="text-[10px] text-secondary">{item.name}</span>
+              <span className={`text-[10px] font-mono ${indexPctClass(pct)}`}>{fmtIndexPct(pct)}</span>
+            </div>
+            <div className={`mt-0.5 truncate font-mono text-[10px] ${indexPctClass(pct)}`}>
+              {fmtIndexValue(value)}
+            </div>
+          </NavLink>
+        )
+      })}
+    </div>
+  )
+}
